Fix hardcoded staff submit bugs
Due to the API not returning staff Recursers with a complete 'stints'
property--i.e. the array containing information about their batches--
the frontend was unable to properly store them in the 'updated_niceties'
Set (it's encoded with the most recent end_date pulled from the 'stints'
array).

So, for the time being, until the API can be changed to work with staffers,
hardcode the 'end_date' for any person with an empty 'stints' property to
Sep 22.