Correctly populate user's image during first login
In commit 7299afee2dc14796e54e9faeaf90b6b2fd9e495d, we started the
migration to the new RC API. That commit correctly updated to use the
new `image_path` key for other Recursers, and the app continues to work
for people who have previously signed in, but one spot was missed: the
initial avatar population for a user signing in for the first time. As a
result, new users were not able to sign in.

Fix the bug by using the correct key of the profile API.

Issue #13 Upgrade to new version of the RC API
PR #30 new RC API calls