Fix is_faculty bug on login
The new RC profiles API does not include an `is_faculty` key, as the old
one did. When we migrated to the new API in commit
7299afee2dc14796e54e9faeaf90b6b2fd9e495d, logging in appears to have
broken: users with existing sessions could continue to log in, but
users without a session could not, and instead got a 500 Internal Server
Error.

Fix the bug by extracting the "does this profile have an employment
stint?" check out into a helper method, and then calling that helper
method during the OAuth flow.

As part of the method extraction refactoring, move the assignment of our
in_faculty key to the format method, and drop the call to the API for
each individual faculty - the profiles API call for all faculty should
return full profiles, and the individual profile API call adds no
additional data. This should speed things up and make the code a bit
simpler to read.

Issue #13 Upgrade to new version of the RC API
PR #30 new RC API calls