switched to using store2 to handle localStorage stringification
Per issue #31, the front-end has been storing entered niceties in 
browser localStorage, in case the window is closed before saving. This 
requires the values stored be strings, which is awkward for other types. 
This patch migrates to using store2 to manage this stringification, 
removing the potential for string bools or nulls to be passed to the 
backend.