Remove empty componentDidMount from <Person>
The componentDidMount function[1] is part of the React lifecycle:

    componentDidMount() is invoked immediately after a component is
    mounted (inserted into the tree). Initialization that requires DOM
    nodes should go here.

Once upon a time, <Person> had some autosave functionality that was
initialized in componentDidMount, but it was removed in commit
a40a1e9015286d6cca1ac171277993036336a2b7.

Delete the now-empty function both for the sake of simplifying the code,
and to reduce the number of changes needed to convert from
`React.createClass` to `React.Component`.

[1] https://reactjs.org/docs/react-component.html#componentdidmount

Issue #18 Upgrade React