Convert <SaveButton> to function component
The `React.createClass` method of creating React components was
deprecated with React 15.5.0[1]. We will need to convert each component
away from using createClass. There were a few alternatives at the time,
including the recommendation they gave then of `class Example extends
React.Component`; however, the current preferred trend is function
components.

Function components were introduced with React 0.14[2], and work great
when the entirety of a component is its render method. (Later versions
of React support function components with state and side effects via
hooks[3], but we're not there yet!)

<SaveButton> is one of our leaf components, which makes it an ideal
component for this kind of conversion: it depends on no other
components, has no state, and has no behavior. Convert it to be a
function component.

As part of this conversion, simplify the logic. We do not need to have
separate branches for whether or not saving is enabled; instead, we can
always specify the onClick callback, pass a boolean value to the
disabled prop, and depend on the button being disabled to prevent the
callback from being invoked. As a result of not needing any branches, we
can write this as an arrow function expression[4].

[1] https://reactjs.org/blog/2017/04/07/react-v15.5.0.html#migrating-from-reactcreateclass
[2] https://reactjs.org/blog/2015/10/07/react-v0.14.html#stateless-function-components
[3] https://reactjs.org/blog/2019/02/06/react-v16.8.0.html
[4] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/Arrow_functions

Issue #18 Upgrade React