Move row calculation out of <NicetyDisplay>
The function that splits a list of niceties into chunks of 4 is not
specific to <NicetyDisplay>; in fact, it can (and probably should) be
replaced with lodash's chunk[1] function. However, for now, simply move
it out of the class and into a standalone function, as a preparatory
step towards converting <NicetyDisplay> into a function component.

[1] https://lodash.com/docs/4.17.15#chunk