Upgrade react-scripts to 0.5.0
Upgrade react-scripts from 0.2.1 to 0.5.0, which is the next version
that has a breaking change which affects us.

react-scripts 0.5.0 changed where it looks for its input
`index.html`[1]:

    You’ll also need to create a new folder called public in the root of
    your project. Then, move index.html and files it references (such as
    a favicon) into that folder.

Changes relating to how files are served do not affect us as we are not
using the server bundled with react-scripts, but are rather using Python
(specifically, gunicorn).

[1] https://github.com/facebook/create-react-app/blob/master/CHANGELOG-0.x.md#moving-indexhtml-into-public-folder

Issue #18 Upgrade React