Convert <NicetyRow> to function component
The `React.createClass` method of creating React components was
deprecated with React 15.5.0[1]. We will need to convert each component
away from using createClass. There were a few alternatives at the time,
including the recommendation they gave then of `class Example extends
React.Component`; however, the current preferred trend is function
components.

Function components were introduced with React 0.14[2], and work great
when the entirety of a component is its render method. (Later versions
of React support function components with state and side effects via
hooks[3], but we're not there yet!)

<NicetyRow> is one of our simpler components, which makes it an ideal
component for this kind of conversion: it has no state, and no behavior.
Convert it to be a function component. Due to its simplicity, it can be
written as an arrow function expression[4].

[1] https://reactjs.org/blog/2017/04/07/react-v15.5.0.html#migrating-from-reactcreateclass
[2] https://reactjs.org/blog/2015/10/07/react-v0.14.html#stateless-function-components
[3] https://reactjs.org/blog/2019/02/06/react-v16.8.0.html
[4] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/Arrow_functions

Issue #18 Upgrade React