Remove hard-coded "special" people
The Niceties app's concept of "special" people is based on an
out-of-date assumption: that we should only show faculty members that
are leaving soon. Since the staff stopped using the Niceties app,
they've been including all faculty members in the Google form.

(It is also based on the assumption that the RC API would not give
departing faculty members' end date; I'm not sure if that is still the
case.)

We want Niceties to always show faculty, and we started doing so in
commit 06019916b489ecba8845b64807fd6b81419f6735. That commit also
commented out the client-side handling of "special" people, so even if
we were populating that field in the API, the client would not show it.

Remove the "special" field in the API, and delete the commented-out
front-end code that used it.