Inline and simplify staff section in <People>
Now that we're always showing faculty members, we can simplify the code
a bit by inlining the variables related to the staff section. At the
same time, we can replace the function expression in the map call with
an arrow function expression, which also lets us avoid a bind call.