Add update_data script and update README
"update_data.py" fetches information from the RC API to populate
the database and have an up-to-date cache. We can schedule data updates
using the Heroku Scheduler.

Update the "README.md" to include instructions on creating an access
token for the RC API requests and set up a Heroku Scheduler.

Issue #68: Normalize Recurser profiles
Issue #5: Clearing the cache in a timely fashion

Co-authored-by: Jason Owen <jason@jasonaowen.net>