Create Profile and Stint tables
We want to maintain a local copy of the RC API data and make our data
model more consistent by including all the relevant information for a
nicety.

Previously, we were relied on making API calls to populate
data we didn't have locally in response to each request to the front
end. We are still doing this, but these data models set the stage to
stop doing that.

Issue #68: Normalize Recurser profiles

Co-authored-by: Jason Owen <jason@jasonaowen.net>