Add nicety.stint_id column + attributes
The nicety.stint_id column will remain nullable until all profiles
are populated with their respective stints in future migrations.

Create a foreign key constraint to tie niceties to stints. Change the
unique constraint to be between the "author_id," "target_id," and
"stint_id." We are still leaving "end_date" alone because we are not
yet cleaning up the backend to use "stints" instead of "end_dates."

Create a foreign key to tie target_ids to profile_ids.

Co-authored-by: Jason Owen <jason@jasonaowen.net>