Create Stints for missing people
The 156 niceties with null stint_ids consist of 18 profiles with 18
end dates.

Out of the 18 profiles, create stints for the newly created profiles
using negative stint ids:

- 9 new profiles manually created (4 Former Recurser's + 5 from
  Users table)
- 2 profiles with no niceties written for them, therefore no need to
  generate stints

This leaves 7 profiles with 8 stints, with 2 stints for one profile

After this, there will be 68 niceties with null stint_ids left.
The subsequent migration will address this.