API Changes (WIP)
These are the changes to the actual backend files to finally pull
data from the database, rather than user RC's API.

This is a draft; the files will need to be organized and old code
will need to be deleted. This push is simply to have the changes
on Github.